# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Space Case Cruises API
  version: 0.2.0
  description: |
    Admin-first community platform API for Space Case Cruises.
    Auth in local/dev uses one of these headers:
    - x-user-email
    - x-user-id
servers:
  - url: http://localhost:4000
security:
  - UserEmailAuth: []
  - UserIdAuth: []
components:
  securitySchemes:
    UserEmailAuth:
      type: apiKey
      in: header
      name: x-user-email
    UserIdAuth:
      type: apiKey
      in: header
      name: x-user-id
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CruiseIdParam:
      name: cruiseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CruiseStatus:
      type: string
      enum: [active, archived]
    VisibilityState:
      type: string
      enum: [invisible, inactive, active]
...
paths:
...
  /api/v1/cruises:
    get:
      tags: [Cruises]
      summary: List cruises
      parameters:
        - in: query
          name: include_archived
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        '200':
          description: Cruises list
  /api/v1/admin/cruises:
    post:
      tags: [Cruises]
      summary: Create cruise (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, year]
              properties:
                name:
                  type: string
                year:
                  type: integer
                  minimum: 2000
                location:
                  type: string
                  nullable: true
                starts_on:
                  type: string
                  format: date
                  nullable: true
                ends_on:
                  type: string
                  format: date
                  nullable: true
                map_image_url:
                  type: string
                  format: uri
                  nullable: true
                is_featured:
                  type: boolean
                sort_order:
                  type: integer
                force_all_subgroups_to_dock:
                  type: boolean
      responses:
        '201':
          description: Created cruise
  /api/v1/admin/cruises/{id}:
    patch:
      tags: [Cruises]
      summary: Update cruise (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                year:
                  type: integer
                  minimum: 2000
                location:
                  type: string
                  nullable: true
                starts_on:
                  type: string
                  format: date
                  nullable: true
                ends_on:
                  type: string
                  format: date
                  nullable: true
                map_image_url:
                  type: string
                  format: uri
                  nullable: true
                status:
                  $ref: '#/components/schemas/CruiseStatus'
                is_featured:
                  type: boolean
                sort_order:
                  type: integer
                force_all_subgroups_to_dock:
                  type: boolean
      responses:
        '200':
          description: Updated cruise
...
  /api/v1/cruises/{cruiseId}/subgroups:
    get:
      tags: [Cruise SubGroups]
      summary: List cruise subgroup assignments
      parameters:
        - $ref: '#/components/parameters/CruiseIdParam'
      responses:
        '200':
          description: Assignment list
  /api/v1/admin/cruises/{cruiseId}/subgroups:
    post:
      tags: [Cruise SubGroups]
      summary: Create cruise subgroup assignment (admin)
      parameters:
        - $ref: '#/components/parameters/CruiseIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [subgroup_id]
              properties:
                subgroup_id:
                  type: string
                  format: uuid
                override_name:
                  type: string
                  nullable: true
                override_description:
                  type: string
                  nullable: true
                detail_image_url:
                  type: string
                  format: uri
                  nullable: true
                cost_level_override:
                  type: integer
                  minimum: 0
                  maximum: 8
                  nullable: true
                visibility_state:
                  $ref: '#/components/schemas/VisibilityState'
                dock_visible:
                  type: boolean
                map_x:
                  type: number
                  minimum: 0
                  maximum: 100
                  nullable: true
                map_y:
                  type: number
                  minimum: 0
                  maximum: 100
                  nullable: true
                map_scale:
                  type: number
                  exclusiveMinimum: 0
                  nullable: true
      responses:
        '201':
          description: Created assignment
  /api/v1/admin/cruise-subgroups/{id}:
    patch:
      tags: [Cruise SubGroups]
      summary: Update cruise subgroup assignment (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                override_name:
                  type: string
                  nullable: true
                override_description:
                  type: string
                  nullable: true
                detail_image_url:
                  type: string
                  format: uri
                  nullable: true
                cost_level_override:
                  type: integer
                  minimum: 0
                  maximum: 8
                  nullable: true
                visibility_state:
                  $ref: '#/components/schemas/VisibilityState'
                dock_visible:
                  type: boolean
                map_x:
                  type: number
                  minimum: 0
                  maximum: 100
                  nullable: true
                map_y:
                  type: number
                  minimum: 0
                  maximum: 100
                  nullable: true
                map_scale:
                  type: number
                  exclusiveMinimum: 0
                  nullable: true
      responses:
        '200':
          description: Updated assignment
...